Windows version builds and works!

+++ Main/Frontend/src/main.js
@@ -3,10 +3,27 @@ const fs = require("fs");
 const path = require("path");
 const process = require("process");
 let mainWindow;
-const filePath = path.join("/tmp/testcases/", "hosts.json");
-fs.rmdir("/tmp/testcases", { recursive: true }, (err) => {
+filePath = ""
+const os = require("os");
+
+const platform = os.platform();
+if (platform === "linux") {
+	fs.rmdir("/tmp/testcases", { recursive: true }, (err) => {
   if (err) console.error(err);
 });
+ filePath = path.join("/tmp/testcases/", "hosts.json");
+}
+else if (platform === "win32") {
+	fs.rmdir("C:\\Windows\\Temp\\testcases", { recursive: true }, (err) => {
+  if (err) console.error(err);
+});
+	filePath = "C:\\Windows\\Temp\\testcases\\hosts.json";
+}
+else {
+	console.log("Err, OS not supported.");
+}
+
+
 
 function createWindow() {
   mainWindow = new BrowserWindow({
@@ -43,7 +60,7 @@ app.whenReady().then(() => {
         mainWindow.webContents.send("json-data", data);
         fileSent = true; // Prevent sending multiple times
       }
-    }, 10000);
+    }, 3000);
     return filePaths[0];
   });
 });
